requirements.txt
openai
google-genai
yt-dlp
youtube-transcript-api
Pillow
python-dotenv
requests
mutagen
google-api-python-client
google-auth-oauthlib
rembg
onnxruntime
